Log parse errors for SettingsFiles as well as for SettingsAutoSave instances.
