+++ Eutherion/Win/Storage/SettingCopy.cs
@@ -20,6 +20,7 @@
 #endregion
 
 using Eutherion.Text.Json;
+using Eutherion.Utils;
 using System;
 using System.Collections.Generic;
 
@@ -183,7 +184,7 @@ namespace Eutherion.Win.Storage
         /// <summary>
         /// Loads settings from text.
         /// </summary>
-        internal List<JsonErrorInfo> TryLoadFromText(string json)
+        internal bool TryLoadFromText(string json)
         {
             var settingSyntaxTree = SettingSyntaxTree.ParseSettings(json, Schema);
 
@@ -199,7 +200,27 @@ namespace Eutherion.Win.Storage
                 }
             }
 
-            return settingSyntaxTree.Errors;
+            // Log parse errors. Return true if no errors were found.
+            var errors = settingSyntaxTree.Errors;
+            if (errors.Count > 0)
+            {
+                errors.ForEach(x => new SettingsParseException(x).Trace());
+                return false;
+            }
+
+            return true;
         }
+    }
+
+    internal class SettingsParseException : Exception
+    {
+        public static string AutoSaveFileParseMessage(JsonErrorInfo jsonErrorInfo)
+        {
+            string paramDisplayString = StringUtilities.ToDefaultParameterListDisplayString(jsonErrorInfo.Parameters);
+            return $"{jsonErrorInfo.ErrorCode}{paramDisplayString} at position {jsonErrorInfo.Start}, length {jsonErrorInfo.Length}";
+        }
+
+        public SettingsParseException(JsonErrorInfo jsonErrorInfo)
+            : base(AutoSaveFileParseMessage(jsonErrorInfo)) { }
     }
 }
